// src/shared/types/payload-types.ts
export interface Product {
  id: number;
  article1C?: string | null;
  title: string;
  slug: string;
  price?: number | null;
  description?: string | null;
  gallery?:
    | {
        image?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  quantity?: number | null;
  category?: (number | null) | Category;
  author: number | Author;
  createdAt: string;
  updatedAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  createdAt: string;
  updatedAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  label: string;
  value: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "authors".
 */
export interface Author {
  id: number;
  name?: string | null;
  fullName?: string | null;
  slug?: string | null;
  bio?: string | null;
  avatar?: (number | null) | Media;
  cover?: (number | null) | Media;
  externalLink?: string | null;
  products_count?: number | null;
  product_categories?:
    | {
        category?: string | null;
        id?: string | null;
      }[]
    | null;
  user?: (number | null) | User;
  createdAt: string;
  updatedAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: number;
  role?: ('admin' | 'author' | 'customer') | null;
  createdAt: string;
  updatedAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
// ...
export interface Cart {
  id: number;
  owner: number | Customer;
  items?:
    | {
        product: number | Product;
        quantity: number;
        checked?: boolean | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "customers".
 */
export interface Customer {
  id: number;
  email: string;
  user: number | User;
  fullName?: string | null;
  phone?: string | null;
  addresses?:
    | {
        label?: string | null;
        addressLine: string;
        city: string;
        postalCode: string;
        id?: string | null;
      }[]
    | null;
  cart?: (number | null) | Cart;
  createdAt: string;
  updatedAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "orders".
 */
export interface Order {
  id: number;
  orderNumber?: string | null;
  customer: number | Customer;
  items: {
    productSnapshot: {
      productId: number;
      title: string;
      price: number;
    };
    quantity: number;
    id?: string | null;
  }[];
  deliveryType: 'pickup' | 'delivery';
  cdekData?: {
    type?: ('pvz' | 'courier') | null;
    address: string;
    code?: string | null;
  };
  customerDetails?: string | null;
  /**
   * Заполняется после отправки заказа
   */
  trackingNumber?: string | null;
  comment?: string | null;
  status: 'prepared' | 'processing' | 'assembled' | 'sent' | 'delivered' | 'completed' | 'cancelled';
  /**
   * Статус обновляется периодически. Актуальное значение можно узнать в ЮКассе
   */
  paymentStatus?: ('pending' | 'waiting_for_capture' | 'succeeded' | 'canceled') | null;
  paymentId?: string | null;
  paymentLink?: string | null;
  total: number;
  createdAt: string;
  updatedAt: string;
}